Fix size_t to int conversion compile error.
BUG=

TBR=mstarzinger@chromium.org

Review URL: https://codereview.chromium.org/552313002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@23805 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/js-inlining.cc
@@ -171,8 +171,8 @@ class CopyVisitor : public NullNodeVisitor {
 
     // Reuse the operator in the copy. This assumes that op lives in a zone
     // that lives longer than graph()'s zone.
-    Node* copy =
-        target_graph_->NewNode(original->op(), inputs.size(), &inputs.front());
+    Node* copy = target_graph_->NewNode(
+        original->op(), static_cast<int>(inputs.size()), &inputs.front());
     copies_[original->id()] = copy;
     return GenericGraphVisit::CONTINUE;
   }
